Meaningful user service names
User provided services contained in files like "beach-service-foo.sh" are now run as Supervisor services with the same name.

Previously, the name of Supervisor services was beach-user-1, beach-user-2 and so forth, and it was difficult to find out which service or process belonged to which beach-service-script. Now the original filename of the service script is used as a service name and can be easily discovered with `supervisorctl status`.