Fix fix for empty FLOW_CONTEXT in Local Beach
In (recent) Local Beach projects, `FLOW_CONTEXT` is set to an empty
value if not given in the environment. This lead to `FLOW_CONTEXT`
being empty, breaking configuration of Flow.

This fixes replacament of the check for being set with a check for
not being empty to fix this by also checking for the variable to be
set.

Otherwise it resulted in `FLOW_CONTEXT: unbound variable` during
startup.